requirements.txt
diff-match-patch==20121119
Django==1.9.7
django-import-export==0.4.5
django-jet==0.1.4
django-report-builder==3.3.0
django-report-utils==0.3.15
djangorestframework==3.3.3
jdcal==1.2
openpyxl==2.2.1
python-dateutil==2.5.3
six==1.10.0
tablib==0.11.2
